Use mount feature for pages
