+++ extension-manifest-v2/app/templates/autoconsent.html
@@ -10,7 +10,5 @@
     <script src="../../dist/autoconsent.js" async></script>
     <link rel="stylesheet" href="../../dist/css/autoconsent.css" />
   </head>
-  <body>
-    <gh-autoconsent></gh-autoconsent>
-  </body>
+  <body></body>
 </html>
