+++ extension-manifest-v2/app/templates/onboarding.html
@@ -21,7 +21,5 @@
 		<script src="../../dist/onboarding.js" async type="module"></script>
 		<link rel="stylesheet" href="../../dist/css/onboarding.css" />
 	</head>
-	<body>
-		<gh-onboarding></gh-onboarding>
-	</body>
+	<body></body>
 </html>
