+++ extension-manifest-v2/app/templates/trackers-preview.html
@@ -11,7 +11,5 @@
     <script src="../../dist/trackers_preview.js" async></script>
     <link rel="stylesheet" href="../../dist/css/trackers_preview.css" />
   </head>
-  <body>
-    <gh-trackers-preview></gh-trackers-preview>
-  </body>
+  <body></body>
 </html>
