+++ extension-manifest-v3/src/pages/autoconsent/index.html
@@ -6,7 +6,5 @@
     <title>Never-Consent</title>
     <script type="module" src="./index.js" async></script>
   </head>
-  <body>
-    <gh-autoconsent></gh-autoconsent>
-  </body>
+  <body></body>
 </html>
