+++ extension-manifest-v3/src/pages/onboarding/iframe.html
@@ -6,7 +6,5 @@
     <title>Ghostery is disabled</title>
     <script src="./iframe.js" type="module" async></script>
   </head>
-  <body>
-    <gh-onboarding-iframe></gh-onboarding-iframe>
-  </body>
+  <body></body>
 </html>
