+++ extension-manifest-v3/src/pages/onboarding/index.html
@@ -7,7 +7,5 @@
     <script src="./index.js" type="module" async></script>
     <link rel="stylesheet" href="./styles.css" />
   </head>
-  <body>
-    <gh-onboarding></gh-onboarding>
-  </body>
+  <body></body>
 </html>
