+++ extension-manifest-v3/src/pages/onboarding/styles.css
@@ -2,8 +2,3 @@ html,
 body {
   height: 100%;
 }
-
-gh-onboarding {
-  display: block;
-  height: 100%;
-}
