+++ extension-manifest-v3/src/pages/panel/index.html
@@ -6,7 +6,5 @@
     <title>Ghostery panel</title>
     <script type="module" src="./index.js" async></script>
   </head>
-  <body tabindex="0">
-    <gh-panel-main></gh-panel-main>
-  </body>
+  <body tabindex="0"></body>
 </html>
