--- extension-manifest-v3/src/pages/panel/views/main.js
@@ -1,21 +1,0 @@
-/**
- * Ghostery Browser Extension
- * https://www.ghostery.com/
- *
- * Copyright 2017-present Ghostery GmbH. All rights reserved.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0
- */
-
-import { html, router } from 'hybrids';
-
-import Home from './home.js';
-
-export default {
-  stack: router([Home]),
-  content: ({ stack }) => html`
-    <template layout="block width:full:350px">${stack}</template>
-  `,
-};
