+++ extension-manifest-v3/src/pages/settings/index.html
@@ -6,7 +6,5 @@
     <title>Ghostery settings</title>
     <script type="module" src="./index.js" async></script>
   </head>
-  <body>
-    <gh-settings-main></gh-settings-main>
-  </body>
+  <body></body>
 </html>
