+++ extension-manifest-v3/src/pages/trackers-preview/index.html
@@ -6,7 +6,5 @@
     <title>Trackers Preview</title>
     <script src="./index.js" type="module" async></script>
   </head>
-  <body>
-    <gh-trackers-preview></gh-trackers-preview>
-  </body>
+  <body></body>
 </html>
